parsers/edi_parser.py
from bots.botslib import BotsError
from bots.botslib import botsinit
from bots.botslib import botsengine as be
from bots.botslib import Job
import os
from typing import List, Dict

def parse_edi(file_path: str) -> List[Dict]:
    """Обработка EDI файлов через Bots"""
    try:
        # Инициализация Bots
        botsinit.generalinit('config')
        
        # Создание задачи обработки
        job = Job()
        job.fromta = {'filename': file_path}
        job.script = 'edifact_in'  # Пример для EDIFACT
        
        # Запуск обработки
        be.runjob(job)
        
        # Получение результатов (упрощенно)
        # В реальном проекте нужно парсить выходные данные Bots
        return [{"edi_data": "processed"}]  # Заглушка
    except BotsError as e:
        raise Exception(f"EDI processing error: {str(e)}")